Consolidate all the separate profile-startup tests into one file.
